pyproject.toml
[tool.poetry]
name = "yufuquant"
version = "0.1.0"
description = ""
authors = ["zmrenwu <zmrenwu@163.com>"]

[tool.poetry.dependencies]
python = "^3.7"
ccxt = "^1.30.52"
django = "^3.0.7"
django-model-utils = "^4.0.0"
djangorestframework = "^3.11.0"
djangorestframework-bulk = "^0.2.1"
django-cryptography = "^1.0"
django-extensions = "^3.0.1"
django-allauth = "^0.42.0"
django-rest-auth = "^0.9.5"
channels = "^2.4.0"
channels-redis = "^2.4.2"
django-cors-headers = "^3.4.0"
django-filter = "^2.3.0"
uvicorn = "^0.11.5"
django-environ = "^0.4.5"
whitenoise = "^5.1.0"
django-imagekit = "^4.0.2"
pillow = "^7.2.0"
gunicorn = "^20.0.4"
sentry-sdk = "^0.15.1"
django-webpack-loader = "^0.7.0"
factory-boy = "^2.12.0"
jsonfield = "^3.1.0"

[tool.poetry.dev-dependencies]
mkdocs = "^1.1.2"
django-debug-toolbar = "^2.2"
pytest = "^5.4.3"
pytest-django = "^3.9.0"
pytest-cov = "^2.10.0"
freezegun = "^0.3.15"
django-silk = "^4.0.1"
django-test-plus = "^1.4.0"
pytest-sugar = "^0.9.3"
django-stubs = "^1.5.0"
djangorestframework-stubs = "^1.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
